refactor: organize ranking queries
